change to long/double (fix implicit type conversions)

+++ src/at/ac/univie/clustering/data/DataWorker.java
@@ -36,6 +36,6 @@ public interface DataWorker {
     /**
      * @return
      */
-    float[] readTuple() throws IOException;
+    double[] readTuple() throws IOException;
 
 }
